requirements.txt
pyarrow
pandas 
dbt-duckdb
tqdm